Replace old ColSpec with new ColSpec
