+++ src/TableTransforms.jl
@@ -21,8 +21,7 @@ import Distributions: quantile, cdf
 
 include("assertions.jl")
 include("distributions.jl")
-# include("colspec.jl")
-include("newcolspec.jl")
+include("colspec.jl")
 include("transforms.jl")
 
 export
