+++ test/colspec.jl
@@ -3,91 +3,106 @@
   tupnames = (:a, :b, :c, :d, :e, :f)
 
   # vector of symbols
-  colspec = [:a, :c, :e]
+  colspec = TableTransforms.colspec([:a, :c, :e])
   snames = TableTransforms.choose(colspec, vecnames)
   @test snames == [:a, :c, :e]
   snames = TableTransforms.choose(colspec, tupnames)
   @test snames == [:a, :c, :e]
 
   # tuple of symbols
-  colspec = (:a, :c, :e)
+  colspec = TableTransforms.colspec((:a, :c, :e))
   snames = TableTransforms.choose(colspec, vecnames)
   @test snames == [:a, :c, :e]
   snames = TableTransforms.choose(colspec, tupnames)
   @test snames == [:a, :c, :e]
 
   # vector of strings
-  colspec = ["a", "c", "e"]
+  colspec = TableTransforms.colspec(["a", "c", "e"])
   snames = TableTransforms.choose(colspec, vecnames)
   @test snames == [:a, :c, :e]
   snames = TableTransforms.choose(colspec, tupnames)
   @test snames == [:a, :c, :e]
 
   # tuple of strings
-  colspec = ("a", "c", "e")
+  colspec = TableTransforms.colspec(("a", "c", "e"))
   snames = TableTransforms.choose(colspec, vecnames)
   @test snames == [:a, :c, :e]
   snames = TableTransforms.choose(colspec, tupnames)
   @test snames == [:a, :c, :e]
 
   # vector of integers
-  colspec = [1, 3, 5]
+  colspec = TableTransforms.colspec([1, 3, 5])
   snames = TableTransforms.choose(colspec, vecnames)
   @test snames == [:a, :c, :e]
   snames = TableTransforms.choose(colspec, tupnames)
   @test snames == [:a, :c, :e]
 
   # tuple of integers
-  colspec = (1, 3, 5)
+  colspec = TableTransforms.colspec((1, 3, 5))
   snames = TableTransforms.choose(colspec, vecnames)
   @test snames == [:a, :c, :e]
   snames = TableTransforms.choose(colspec, tupnames)
   @test snames == [:a, :c, :e]
 
   # regex
-  colspec = r"[ace]"
+  colspec = TableTransforms.colspec(r"[ace]")
   snames = TableTransforms.choose(colspec, vecnames)
   @test snames == [:a, :c, :e]
   snames = TableTransforms.choose(colspec, tupnames)
   @test snames == [:a, :c, :e]
 
   # colon
-  snames = TableTransforms.choose(:, vecnames)
+  colspec = TableTransforms.colspec(:)
+  snames = TableTransforms.choose(colspec, vecnames)
   @test snames == [:a, :b, :c, :d, :e, :f]
-  snames = TableTransforms.choose(:, tupnames)
+  snames = TableTransforms.choose(colspec, tupnames)
   @test snames == [:a, :b, :c, :d, :e, :f]
 
   # nothing
-  snames = TableTransforms.choose(nothing, vecnames)
+  colspec = TableTransforms.colspec(nothing)
+  snames = TableTransforms.choose(colspec, vecnames)
   @test snames == Symbol[]
-  snames = TableTransforms.choose(nothing, tupnames)
+  snames = TableTransforms.choose(colspec, tupnames)
   @test snames == Symbol[]
 
   # throws
-  @test_throws AssertionError TableTransforms.choose(r"x", vecnames)
-  @test_throws AssertionError TableTransforms.choose(r"x", tupnames)
-  @test_throws AssertionError TableTransforms.choose(String[], vecnames)
-  @test_throws AssertionError TableTransforms.choose(String[], tupnames)
-  @test_throws AssertionError TableTransforms.choose(Symbol[], vecnames)
-  @test_throws AssertionError TableTransforms.choose(Symbol[], tupnames)
+  colspec = TableTransforms.colspec(r"x")
+  @test_throws AssertionError TableTransforms.choose(colspec, vecnames)
+  @test_throws AssertionError TableTransforms.choose(colspec, tupnames)
+  @test_throws AssertionError TableTransforms.colspec(Symbol[])
+  @test_throws AssertionError TableTransforms.colspec(String[])
+  @test_throws AssertionError TableTransforms.colspec(Int[])
+  @test_throws ArgumentError TableTransforms.colspec(())
+  @test_throws ArgumentError TableTransforms.colspec(missing)
 
   # type stability
-  @inferred TableTransforms.choose([:a, :b], vecnames)
-  @inferred TableTransforms.choose([:a, :b], tupnames)
-  @inferred TableTransforms.choose((:a, :b), vecnames)
-  @inferred TableTransforms.choose((:a, :b), tupnames)
-  @inferred TableTransforms.choose(["a", "b"], vecnames)
-  @inferred TableTransforms.choose(["a", "b"], tupnames)
-  @inferred TableTransforms.choose(("a", "b"), vecnames)
-  @inferred TableTransforms.choose(("a", "b"), tupnames)
-  @inferred TableTransforms.choose([1, 2], vecnames)
-  @inferred TableTransforms.choose([1, 2], tupnames)
-  @inferred TableTransforms.choose((1, 2), vecnames)
-  @inferred TableTransforms.choose((1, 2), tupnames)
-  @inferred TableTransforms.choose(r"[ab]", vecnames)
-  @inferred TableTransforms.choose(r"[ab]", tupnames)
-  @inferred TableTransforms.choose(:, vecnames)
-  @inferred TableTransforms.choose(:, tupnames)
+  colspec = TableTransforms.colspec([:a, :b])
+  @inferred TableTransforms.choose(colspec, vecnames)
+  @inferred TableTransforms.choose(colspec, tupnames)
+  colspec = TableTransforms.colspec((:a, :b))
+  @inferred TableTransforms.choose(colspec, vecnames)
+  @inferred TableTransforms.choose(colspec, tupnames)
+  colspec = TableTransforms.colspec(["a", "b"])
+  @inferred TableTransforms.choose(colspec, vecnames)
+  @inferred TableTransforms.choose(colspec, tupnames)
+  colspec = TableTransforms.colspec(("a", "b"))
+  @inferred TableTransforms.choose(colspec, vecnames)
+  @inferred TableTransforms.choose(colspec, tupnames)
+  colspec = TableTransforms.colspec([1, 2])
+  @inferred TableTransforms.choose(colspec, vecnames)
+  @inferred TableTransforms.choose(colspec, tupnames)
+  colspec = TableTransforms.colspec((1, 2))
+  @inferred TableTransforms.choose(colspec, vecnames)
+  @inferred TableTransforms.choose(colspec, tupnames)
+  colspec = TableTransforms.colspec(r"[ab]")
+  @inferred TableTransforms.choose(colspec, vecnames)
+  @inferred TableTransforms.choose(colspec, tupnames)
+  colspec = TableTransforms.colspec(:)
+  @inferred TableTransforms.choose(colspec, vecnames)
+  @inferred TableTransforms.choose(colspec, tupnames)
+  colspec = TableTransforms.colspec(nothing)
+  @inferred TableTransforms.choose(colspec, vecnames)
+  @inferred TableTransforms.choose(colspec, tupnames)
 end
 
 @testset "TableSelection" begin
